# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Health App
  version: "1.0"
  description: |-
    My health app. Right now only supports tracking workouts.
servers:
  - url: https://health-app.com/api/v1
  - url: localhost:8080/api/v1
...
paths:
...
  /workout/user/{userId}:
    get:
      tags:
        - workout
      summary: get all workouts for user given by the param userId
      operationId: getWorkoutByUserId
      parameters:
        - name: userId
          in: path
          description: the id of the current user
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workout"

    post:
      tags:
        - workout
      summary: create a new workout for user given by the param userId
      operationId: createWorkout
      parameters:
        - name: userId
          in: path
          description: the id of the current user
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workout'
      responses:
        '200':
          description: Success
...
components:
  schemas:
...
    Workout:
      type: object
      properties:
        workoutId:
          type: integer
          format: int32
        userId:
          type: string
          format: uuid
        name:
          type: string
        completionDate:
          type: string
          format: date-time
        setExerciseGroups:
          type: array
          items:
              $ref: '#/components/schemas/SetExerciseGroup'
    SetExerciseGroup:
      type: object
      properties:
        setGroupId:
          type: integer
          format: int32
        exercise:
          type: string
          example: "squat"
        totalSets:
          type: integer
          format: int32
          example: 3
        repetitions:
          type: array
          items:
            type: integer
            format: int32
          example: [10, 8, 8]
        weights:
          type: array
          items:
            type: number
            format: float
            example: 22.5
          example: [22.5, 25, 25]
        unit:
          type: string
          example: "pound"
        unitAbbreviation:
          type: string
          example: "lb"
        restPeriods:
          type: array
          description: rest period in seconds
          items:
            type: integer
            format: int32
          example: [60, 90, 60]
        intensities:
          type: array
          items:
            type: integer
            format: int32
          example: [7, 8, 8]
          description: ratings of perceived exertion of each set. Scale of 1 to 10.
        notes:
          type: string
          description: Any notes about how this set exercise set group
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header